pyproject.toml
[tool.poetry]
name = "game-master"
version = "0.1.0"
description = "agent that runs games"
authors = ["Richard Bakobiibizo <110358293+Bakobiibizo@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
openai = "^0.28.1"
anthropic = "^0.3.11"
sentence-transformers = "^2.2.2"
python-cassandra = "^0.6.1"
cassandra-driver = "^3.28.0"
langchain = "^0.0.306"
tiktoken = "^0.5.1"
cassio = "^0.1.3"
pandas = "^2.1.1"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
